+++ app/src/main/java/com/muse1991/museapp/SignInFragment.kt
@@ -1,0 +1,77 @@
+package com.muse1991.museapp
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.GoogleAuthProvider
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import com.muse1991.museapp.databinding.FragmentSignInBinding
+
+class SignInFragment : Fragment() {
+    private lateinit var binding: FragmentSignInBinding
+    private lateinit var auth: FirebaseAuth
+    private lateinit var signInClient: GoogleSignInClient
+    private lateinit var activityResultLauncher: ActivityResultLauncher<Intent>
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = DataBindingUtil.inflate(
+            layoutInflater, R.layout.fragment_sign_in, container, false
+        )
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        setSignInClient()
+        setSignInButton()
+        registerActivityResult()
+    }
+
+    private fun setSignInClient() {
+        val signInOptions = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+            .requestIdToken(getString(R.string.web_client_id))
+            .requestEmail()
+            .build()
+        signInClient = GoogleSignIn.getClient(requireActivity(), signInOptions)
+        auth = Firebase.auth
+    }
+
+    private fun setSignInButton() {
+        binding.imageViewSignIn.setOnClickListener {
+            val signInIntent = signInClient.signInIntent
+            activityResultLauncher.launch(signInIntent)
+        }
+    }
+
+    private fun registerActivityResult() {
+        activityResultLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
+            if (result.resultCode == AppCompatActivity.RESULT_OK) {
+                val idToken = GoogleSignIn.getSignedInAccountFromIntent(result.data).result.idToken
+                val credential = GoogleAuthProvider.getCredential(idToken, null)
+
+                auth.signInWithCredential(credential)
+                    .addOnCompleteListener(requireActivity()) { task ->
+                        if (task.isSuccessful) {
+                            //TODO 화면 이동
+                        }
+                    }
+            }
+        }
+    }
+}
